Add base entity to sabnzbd
