+++ homeassistant/components/sabnzbd/sensor.py
@@ -13,13 +13,12 @@ from homeassistant.components.sensor import (
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import UnitOfDataRate, UnitOfInformation
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import StateType
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from . import DOMAIN, SabnzbdUpdateCoordinator
-from .const import DEFAULT_NAME
+from .const import DOMAIN
+from .coordinator import SabnzbdUpdateCoordinator
+from .entity import SabnzbdEntity
 
 
 @dataclass(frozen=True, kw_only=True)
@@ -139,34 +138,13 @@ async def async_setup_entry(
     entry_id = config_entry.entry_id
     coordinator: SabnzbdUpdateCoordinator = hass.data[DOMAIN][entry_id]
 
-    async_add_entities(
-        [SabnzbdSensor(coordinator, sensor, entry_id) for sensor in SENSOR_TYPES]
-    )
+    async_add_entities([SabnzbdSensor(coordinator, sensor) for sensor in SENSOR_TYPES])
 
 
-class SabnzbdSensor(CoordinatorEntity[SabnzbdUpdateCoordinator], SensorEntity):
+class SabnzbdSensor(SabnzbdEntity, SensorEntity):
     """Representation of an SABnzbd sensor."""
 
     entity_description: SabnzbdSensorEntityDescription
-    _attr_should_poll = False
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        coordinator: SabnzbdUpdateCoordinator,
-        description: SabnzbdSensorEntityDescription,
-        entry_id,
-    ) -> None:
-        """Initialize the sensor."""
-        super().__init__(coordinator)
-
-        self._attr_unique_id = f"{entry_id}_{description.key}"
-        self.entity_description = description
-        self._attr_device_info = DeviceInfo(
-            entry_type=DeviceEntryType.SERVICE,
-            identifiers={(DOMAIN, entry_id)},
-            name=DEFAULT_NAME,
-        )
 
     @property
     def native_value(self) -> StateType:
